requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.0
logging==0.4.9.6
pathlib==1.0.1
typing-extensions==4.8.0 